+++ lib/l10n/app_localizations_ja.dart
@@ -2111,6 +2111,9 @@ class AppLocalizationsJa extends AppLocalizations {
   String get overdue => 'Overdue';
 
   @override
+  String get addTopic => 'Add topic';
+
+  @override
   String get accessTokens => 'Access Tokens';
 
   @override