+++ lib/l10n/app_localizations_pt.dart
@@ -2124,6 +2124,9 @@ class AppLocalizationsPt extends AppLocalizations {
   String get overdue => 'Overdue';
 
   @override
+  String get addTopic => 'Add topic';
+
+  @override
   String get accessTokens => 'Access Tokens';
 
   @override